fix(tests): Use correct temporary directory for signature tests
This resolves the issue of a race condition where one test
deletes a temporary directory another test is using.
To be concrete, the resulting RPM of test case
`test_signature` was stored in the same temporary directory as
the one of `test_not_compressed` which may or may not be deleted at the
time of RPM creation.